Added likes counter, added delete button for jokes,
Now when a user likes a joke, the app will check if the joke has already been liked before. If the joke has never been liked before it will save the joke in the database with a like counter of 1, if the joke has been liked before by other users, then instead of saving another joke in the database, the app will only increase the like counter by 1.